requirements.txt
tweepy
python-dotenv
openai
black
numpy
psycopg2-binary
pytz
friendtech
web3
schedule
